+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [23.5] - 2023-10-04
 ### Changed
 - Restart the broker when its SSLListener secret changes. See
   [#317](https://github.com/astarte-platform/astarte-kubernetes-operator/issues/317).